fix(ci): remove redundant u32 cast flagged by clippy under rustc 1.98.0
CI's clippy -D warnings failed on unnecessary_cast: author_posts.len()
already returns u32, so the trailing `as u32` was a no-op that only
newer clippy started flagging.

+++ packages/contracts/contracts/linkora-contracts/src/lib.rs
@@ -1104,7 +1104,7 @@ impl LinkoraContract {
             Self::bump(&env, &author_key);
         }
 
-        let remaining_entries: u32 = f_count + following_count + author_posts.len() as u32;
+        let remaining_entries: u32 = f_count + following_count + author_posts.len();
 
         if f_count == 0 && following_count == 0 && author_posts.is_empty() {
             env.storage().persistent().remove(&tombstone_key);
